made class hierarchy more logical
