+++ app/src/main/java/com/kitrady/PatternRunner.java
@@ -6,19 +6,13 @@ public class PatternRunner {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
 
-        InputGetter objInputGetter = new InputGetter(1.2, 4, 4);
+        InputHandler objInputHandler = new InputHandler(1.2, 4, 4);
 //        InputGetter objInputGetter = new InputGetter(input);
-        InputConverter objInputConverter = new InputConverter(objInputGetter.getRadius(), objInputGetter.getGauge(), objInputGetter.getVertGauge());
-        SphereMaker objSphereMaker = new SphereMaker(objInputConverter.getStRadius(), objInputConverter.getRowCircumference());
+        SphereMaker objSphereMaker = new SphereMaker(objInputHandler.getStRadius(), objInputHandler.getRowCircumference());
 
-        System.out.println("\nobjInputGetter:" + objInputGetter);
-        System.out.println("\nobjInputConverter:" + objInputConverter);
+        System.out.println("\nobjInputGetter:" + objInputHandler);
         System.out.println("\nobjSphereMaker:" + objSphereMaker);
 
         objSphereMaker.printRows();
-
-        // TODO
-        //  - make it properly handle middle row (aka repeat or not)
-        //  - also make handle input methods be one method that takes messages as parameters
     }
 }
